:toolbox: relocate driver module
